requirements.txt
Django==4.2
djangorestframework==3.14
joblib==1.3.0
scikit-learn==1.3.2
pandas==2.0
numpy==1.25.0
Pillow==10.0.0
django-cors-headers==3.14.0
